Cleaned up and created a component for create team component
